# src/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Simple API",
        "description": "A simple API to learn how to write OpenAPI Specification"
    },
    "schemes": [
        "https"
    ],
    "host": "simple.api",
    "basePath": "/openapi101",
    "paths": {
        "/": {
            "get": {
                "summary": "Server status",
                "description": "Returns if the server 2-DAMVI is up",
                "responses": {
                    "200": {
                        "description": "message saying the server is up",
                        "schema": {
                            "type": "array",
                            "items": {
                                "properties": {
                                    "code": {
                                        "type": "string",
                                        "description": "error code"
                                    },
                                    "error": {
                                        "type": "boolean",
                                        "description":
                                        "true means there is an error"
                                    },
                                    "message": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/ranking": {
            "get": {
                "summary": "Players ranking",
                "description": "Returns a list containing all the players ordered by score.",
                "responses": {
                    "200": {
                        "description": "A list of Players",
                        "schema": {
                            "type": "array",
                            "items": {
                                "properties": {
                                    "position": {
                                        "type": "string",
                                        "description": "position on the ranking"
                                    },
                                    "alias": {
                                        "type": "string",
                                        "description": "alias of the player"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "surname":{
                                        "type": "string"
                                    },
                                    "score":{
                                        "type": "integer"
                                    },
                                    "created":{
                                        "type": "string",
                                        "format": "date",
                                        "description": "Date when the player was created"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/players/:alias": {
            "post": {
                "summary": "Player creation",
                "description": "Returns an error message which says if the player searched by alias exists or not.",
                "parameters": [{
                    "in": "path",
                    "name": "alias",
                    "required" : true,
                    "description": "alias of a not existing player you want to save",
                    "type": "string"
                } ],
                "responses": {
                    "201": {
                        "description": "Returns the answer to the search",
                        "schema": {
                            "type": "array",
                            "items": {
                                "properties": {
                                    "code":{
                                        "type": "string",
                                        "description": "error code"
                                    },
                                    "error": {
                                        "type": "boolean",
                                        "description":
                                        "true means there is an error"
                                    },
                                    "message": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            },"put": {
                "summary": "Player update",
                "description": "Returns an error message which says if the player searched by alias already exists or not, also says if parameters are inserted correctly.",
                "parameters": [{
                    "in": "path",
                    "name": "alias",
                    "required" : true,
                    "description": "alias of a not existing player you want to update",
                    "type": "string"
                } ],
                "responses": {
                    "201": {
                        "description": "Returns the answer to the update",
                        "schema": {
                            "type": "array",
                            "items": {
                                "properties": {
                                    "code":{
                                        "type": "string",
                                        "description": "error code"
                                    },
                                    "error": {
                                        "type": "boolean",
                                        "description":
                                        "true means there is an error"
                                    },
                                    "message": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}
        
            